requirements.txt
streamlit
langchain
langchain-groq
langchain-community
chromadb
sentence-transformers
beautifulsoup4
requests
python-dotenv
huggingface-hub